+++ src/features/optional/major/betterDevilAngelRooms/callbacks/preGetCollectible.ts
@@ -8,6 +8,7 @@ import {
 import {
   game,
   getCollectibleName,
+  getCollectibleQuality,
   giveTrinketsBack,
   inAngelShop,
   inGenesisRoom,
@@ -67,7 +68,8 @@ export function betterDevilAngelRoomsPreGetCollectible(
     return CollectibleTypeCustom.DEBUG;
   }
 
-  const collectibleTypeInOrder = getDevilOrAngelItemInOrder(itemPoolType);
+  const collectibleTypeInOrder =
+    getDevilOrAngelCollectibleInOrder(itemPoolType);
   const collectibleName =
     collectibleTypeInOrder === undefined
       ? "Unknown"
@@ -78,10 +80,14 @@ export function betterDevilAngelRoomsPreGetCollectible(
   return collectibleTypeInOrder;
 }
 
-function getDevilOrAngelItemInOrder(
+function getDevilOrAngelCollectibleInOrder(
   itemPoolType: ItemPoolType.DEVIL | ItemPoolType.ANGEL,
 ) {
   const player = Isaac.GetPlayer();
+
+  // The NO trinket has a special effect when it is golden.
+  const numNoTrinket = player.GetTrinketMultiplier(TrinketType.NO);
+  const shouldRerollQuality0 = numNoTrinket >= 2;
 
   // We need to account for the NO trinket; if the player has it, we need to temporarily remove it,
   // otherwise the random items selected will not be consistent.
@@ -90,31 +96,39 @@ function getDevilOrAngelItemInOrder(
   // Only attempt to find a valid item for N iterations in case something goes wrong.
   for (let i = 0; i < MAX_GET_COLLECTIBLE_ATTEMPTS; i++) {
     v.run.gettingCollectible = true;
-    const subType = getNewSubType(itemPoolType);
+    const collectibleType = getNewCollectibleType(itemPoolType);
     v.run.gettingCollectible = false;
 
     // Simply return the new sub-type if we do not have the NO trinket.
     if (trinketSituation === undefined) {
-      return subType;
+      return collectibleType;
     }
 
     // Otherwise, check to see if this is an active item.
-    const itemConfigItem = itemConfig.GetCollectible(subType);
-    if (itemConfigItem === undefined) {
+    const itemConfigItem = itemConfig.GetCollectible(collectibleType);
+    if (
+      itemConfigItem === undefined ||
+      itemConfigItem.Type === ItemType.ACTIVE
+    ) {
       continue;
     }
 
-    if (itemConfigItem.Type !== ItemType.ACTIVE) {
-      // It is not an active item. Give the NO trinket back and return the new sub-type.
-      giveTrinketsBack(player, trinketSituation);
-      return subType;
+    if (shouldRerollQuality0) {
+      const quality = getCollectibleQuality(collectibleType);
+      if (quality === 0) {
+        continue;
+      }
     }
+
+    // It is not an active item. Give the NO trinket back and return the new sub-type.
+    giveTrinketsBack(player, trinketSituation);
+    return collectibleType;
   }
 
   return undefined;
 }
 
-function getNewSubType(
+function getNewCollectibleType(
   itemPoolType: ItemPoolType.DEVIL | ItemPoolType.ANGEL,
 ): CollectibleType {
   switch (itemPoolType) {